executor/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
pyyaml
httpx
pytest
